Fix for debian.
git-svn-id: https://ilk.uvt.nl/svn/trunk/sources/wopr@2473 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ src/runrunrun.cc
@@ -3165,7 +3165,7 @@ int multi( Logfile& l, Config& c ) {
       return -1;
     }
     read_classifiers_from_file( file_kvs, cls );
-    l.log( to_str(cls.size()) );
+    l.log( to_str((int)cls.size()) );
     file_kvs.close();
     for ( cli = cls.begin(); cli != cls.end(); cli++ ) {
       l.log( (*cli)->id );
